02_restful-api-django/pontos_turisticos/requirements-dev.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-3-jet==1.0.8
django-filter==2.4.0
djangorestframework==3.12.1
Pillow==8.0.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
